move command out of index.ts
